requirements.txt
flask
python-Levenshtein
fuzzywuzzy
python-telegram-bot

